Move AddTagsView into it's own file.
