+++ macos/Bookmarks/Views/ContentView.swift
@@ -61,90 +61,6 @@ struct BorderedSelection: ViewModifier {
 
 }
 
-
-struct AddTagsView: View {
-
-    @Environment(\.presentationMode) var presentationMode
-    @Environment(\.manager) var manager: BookmarksManager
-    var items: [Item]
-    @State var isBusy = false
-
-    @State var tokens: [Token<String>] = []  // TODO: Support a payload on the tokens.
-
-    @StateObject var tagsView: TagsView
-    @State var trie = Trie()
-
-    init(database: Database, items: [Item]) {
-        _tagsView = StateObject(wrappedValue: TagsView(database: database))
-        self.items = items
-    }
-
-    var characterSet: CharacterSet {
-        let characterSet = NSTokenField.defaultTokenizingCharacterSet
-        let spaceCharacterSet = CharacterSet(charactersIn: " ")
-        return characterSet.union(spaceCharacterSet)
-    }
-
-    // TODO: Default focus when launching.
-    var body: some View {
-        Form {
-            Section {
-                TokenField("Add tags...", tokens: $tokens) { string in
-                    let tag = string.lowercased()
-                    return Token(tag)
-                        .associatedValue(tag)
-                } completions: { substring in
-                    trie.findWordsWithPrefix(prefix: substring)
-                }
-                .tokenizingCharacterSet(characterSet)
-                .font(.title)
-                .frame(minWidth: 400)
-                HStack {
-                    Spacer()
-                    Button("Cancel") {
-                        presentationMode.wrappedValue.dismiss()
-                    }
-                    .keyboardShortcut(.cancelAction)
-                    Button("OK") {
-                        isBusy = true
-                        let tags = tokens.compactMap { $0.associatedValue }
-                        for item in items {
-                            let item = item
-                                .adding(tags: Set(tags))
-                                .setting(toRead: false)  // TODO: This is a hack that should be removed (maybe make it an option?)
-                            manager.updateItem(item, completion: { _ in })
-                        }
-                        presentationMode.wrappedValue.dismiss()
-                    }
-                    .keyboardShortcut(.defaultAction)
-                }
-            }
-        }
-        .frame(minWidth: 200)
-        .padding()
-        .disabled(isBusy)
-        .onAppear {
-            tagsView.start()
-        }
-        .onDisappear {
-            tagsView.stop()
-        }
-        .onReceive(tagsView.$tags) { tags in
-            DispatchQueue.global(qos: .background).async {
-                let trie = Trie()
-                for tag in tags {
-                    trie.insert(word: tag)
-                }
-                DispatchQueue.main.async {
-                    self.trie = trie
-                }
-            }
-        }
-
-    }
-
-}
-
 extension Item: Hashable {
 
     public func hash(into hasher: inout Hasher) {
